Refactor to be able to reuse UploadAttachmentCommand for other domain obects as Actions, etc.
