+++ src/Equinor.Procosys.Preservation.Command/TagAttachmentCommands/Upload/UploadTagAttachmentCommand.cs
@@ -1,19 +1,15 @@
-﻿using System;
-using System.IO;
-using System.Text.Json.Serialization;
+﻿using System.IO;
 using MediatR;
 using ServiceResult;
 
 namespace Equinor.Procosys.Preservation.Command.TagAttachmentCommands.Upload
 {
-    public class UploadTagAttachmentCommand : IRequest<Result<int>>, ITagCommandRequest, IDisposable
+    public class UploadTagAttachmentCommand : UploadAttachmentCommand, IRequest<Result<int>>, ITagCommandRequest
     {
-        private bool _isDisposed;
-
         public UploadTagAttachmentCommand(int tagId, string fileName, bool overwriteIfExists, Stream content)
+            : base(content)
         {
             TagId = tagId;
-            Content = content ?? throw new ArgumentNullException(nameof(content));
             FileName = fileName;
             OverwriteIfExists = overwriteIfExists;
         }
@@ -21,30 +17,5 @@ namespace Equinor.Procosys.Preservation.Command.TagAttachmentCommands.Upload
         public int TagId { get; }
         public string FileName { get; }
         public bool OverwriteIfExists { get; }
-
-        // JsonIgnore needed here so GlobalExceptionHandler do not try to serialize the Stream when reporting validation errors. 
-        [JsonIgnore]
-        public Stream Content { get; }
-
-        public void Dispose(bool disposing)
-        {
-            if (_isDisposed)
-            {
-                return;
-            }
-
-            if (disposing)
-            {
-                Content.Dispose();
-            }
-       
-            _isDisposed = true;
-        }
-
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
     }
 }
